newblog/requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
autopep8==1.3.2
bcrypt==3.1.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
Django==1.11.4
django-haystack==2.6.1
hyperlink==17.3.0
idna==2.6
incremental==17.5.0
jieba==0.39
lxml==3.8.0
Markdown==2.6.9
paramiko==2.2.1
parsel==1.2.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycodestyle==2.3.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyMySQL==0.7.11
PyNaCl==1.1.2
pyOpenSSL==17.2.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.4
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
urllib3==1.22
w3lib==1.18.0
Whoosh==2.7.4
zope.interface==4.4.2
